LayoutTests/fast/text-extraction/debug-text-extraction-offscreen-button-expected.txt

PASS clickError is ""
PASS clickCount.textContent is "1"
PASS successfullyParsed is true

TEST COMPLETE
Click Me
1
